# backend/docs/swagger.yaml
basePath: /
definitions:
...
  v1.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  v1.UserPatchRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  v1.UserReplaceRequest:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  v1.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: Music streaming backend
  title: Chimera API
  version: "1.0"
paths:
...
  /v1/me:
    delete:
      operationId: deleteCurrentUser
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete current account
      tags:
      - users
    get:
      operationId: getCurrentUser
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - auth
    patch:
      consumes:
      - application/json
      operationId: updateCurrentUser
      parameters:
      - description: Profile fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.UserPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current profile
      tags:
      - users
    put:
      consumes:
      - application/json
      operationId: replaceCurrentUser
      parameters:
      - description: Profile
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.UserReplaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Replace current profile
      tags:
      - users
...
  /v1/users:
    get:
      operationId: listUsers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List users
      tags:
      - users
  /v1/users/{userId}:
    get:
      operationId: getUser
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get user by id
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: JWT from /v1/auth/login. Paste only the token, or "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"